Remove app-misc/tmux overlay and only supply a patch
The app-misc/tmux ebuild already invokes the epatch_user infrastructure
which allows for patches to be added without modification of the actual
ebuild.
Because the only modification we did to the ebuild was the applying of
the patch, this change moves the patch in the respective folder to be
automatically applied while removing the overlay.