Add media-gfx/geeqie patch to support png export
Out of the box, geeqie does not support conversion of image files it
displays. For RAW files, in particular, it applies very nice corrections
automatically not easily reproducibl manually in ufraw or similar
programs.
The added patch adds support for exporting the displayed image in the
form of a png.