Add support for installing documentation to dev-vcs/git-subrepo
With recent changes git-subrepo comes with included documentation in the
form of man pages, html files, and more.
This change adjusts the ebuild to add support for installing those files
in the directories where they can be found by the respective programs.
The installation is guarded by the doc USE flag which is enabled by
default (the rationale being that a 'git help' invocation causes a
search for the man page of the command of interest so by default we want
this man page to be available).