Move app-misc/tmux patches into versioned directory
The malloc trim patch will not apply properly in conjunction with tmux
versions >2.0.
To that end, this change moves the patch (along with the auto rename
fix) into a versioned directory. This way, portage will apply the patch
only for version 1.9a which it was used with. New patches will be
devised for version 2.1 and put into a similar directory.