Remove x11-terms/rxvt-unicode overlay and only rely on patches
The only purpose of the rxvt-unicode overlay was to apply custom patches
to the ebuild. However, the epatch_user feature does an equally well job
without the associated overhead of managing the ebuild itself.
This change removes the entire x11-terms/rxvt-unicode ebuild and moves
all patches into the appropriate folder for them to be applied
correctly.