Remove no longer used x11-terms/rxvt-unicode patches
It has been some time since we used the tabbing capabilities of urxvt in
the form of a Perl extension after emplying tmux. In fact, we have not
enabled the 'perl' USE flag for quite some time.
This change removes four patches associated with the Perl tabbing
functionality. The fifth patch removes enables the scrolling using the
cursor up and down keys -- functionality that is also no longer used.