Add support for installing the git.subrepo module to dev-vcs/git-subrepo
The git-subrepo program not only comprises a script that contains the
functionality in a CLI accessible manner, it also sports a Python module
that can be used when interfacing with it from a Python program. This
change introduces a new USE flag 'module' (enabled by default) that
causes portage to also install the module.