Add mail-mta/msmtp patch to remove overly eager security check
Msmtp seizes to work in case the supplied configuration contains a
secret (i.e., a password) and the file is not owned by the invoking
user.
This restriction is impractical for our use case and there is no obvious
way to work around it. Hence, this change adds a patch that removes the
check.