Add sys-apps/openrc patch to remove timeout warnings
With our dmcrypt setup asking for a passphrase during boot we experience
warnings related to other services that are blocked from starting. These
messages repeatedly warn about a timeout approaching.
These messages are not useful and they destroy the password input
"dialog". This change adds a patch to OpenRC that removes the warning
altogether. This patch is merely a workaround (for lack of a better
solution).