Remove sys-apps/openrc timeout warning removal patch
The patch only removes the warnings while the timeouts internally still
apply. The result is that if an interactive service is not interacted
with in time, for example, the system will go into an undesired state
because some services failed to start.
This change removes the patch. The approach we have to take is to
adjust the services in question to impose no timeout on dependencies.