Use contained setup.py and distutils for dev-python/cleanup ebuild
The world is not build on top of Gentoo's ebuilds. For Python, in
particular, using PyPi or other means of distribution are arguably more
common than system level package managers. The latest version of the
cleanup package provides a setup.py script to install it in a generic
fashion.
This change adjusts the ebuild to make use of setup.py in conjunction
with the distutils wrappers in order to install the package.