Remove mail-client/thunderbird overlay
The mail-client/thunderbird overlay was merely meant to be a way to keep
version 38 of the program as it was removed from the official Gentoo
Portage tree.
By now we moved on to a newer version that is officially supported
again. To that end, this change removes the overlay along with the
eclass files it depends on.