Set custom configuration for www-client/chromium
Quite some time ago when the chromium build silently changed the build
system we were no longer able to set our custom configuration options to
customize the build.
After some investigation we are now back at setting a bunch of
configuration options slimming down the browser a tad (or so is the hope
at least).