Add patchset for x11-wm/i3
This change adds a patchset for the x11-wm/i3 to tailor its behavior to
Our needs. The patches include those for no longer wrapping around when
Changing the focus of windows, fixing the unintuitive resizing logic,
Keeping unused workspaces, and adding an indication which workspaces
Actually contain windows in the bar.