Automatically install completion functionality for media-sound/mpd-complete
So far the ebuild installed the script providing the core completion
functionality but did not actually install something that hooks it up
with the shell in use.
With this change the ebuild is adjusted to install the mpd-complete
script providing completion support for bash in the proper location. In
order for a user to make use of it this script needs to be sourced from
the current shell.