Fix problems arising from PORTAGE_TMPDIR for dev-python/pyvenv-shared
Installing on different systems the following problem was noticed:
the emerge process could fail depending on the value of the
PORTAGE_TMPDIR environment variable. This variable determines the
temporary directory that Portage uses and, hence, influences the
directory in which the emerge happens and the virtual environment we
create resides. The patch we apply, however, assumes an absolute path,
that corresponds to one particular value of PORTAGE_TMPDIR.

To fix this problem, the patch got adjusted to have a dummy path that
gets patched by the ebuild before applying the patch.