Add x11-misc/slim patch to set theme when using auto login feature
This change adds a user patch to x11-misc/slim that causes the program
to properly set the theme when using the auto login feature. Since using
this feature I have noticed that the background was just black and the
Gentoo wallpaper was no longer set. With this change things are back to
"normal".