Adjust dev-lang/rust-source to automatically set RUST_SRC_PATH
So far we have manually set RUST_SRC_PATH (required by programs such as
dev-util/racer) to the directory where the sources were installed. That
is error prone (in case the path changes) and uncomfortable (every user
has to do that).
This change adjusts the ebuild to install a file containing the
respective environment variable in /etc/env.d/ where it is available
globally.