Adjust dev-util/clippy to set RUSTFLAGS properly
Clippy required the nightly compiler for building and running.
Unfortunately, at runtime the compiler is not able to find the
corresponding libraries if the system compiler is set to the stable
version of Rust.
This change adjusts the ebuild to install an env.d file that sets
RUSTFLAGS to add the respective directory in the library search path.