Update x11-themes/slim-theme-gentoo to resize background image
With the recent switch towards using tiling over stretching for the
background image we can see a black margin on the right side of the
screen on most monitors. The problem is that the background image is
only 1600 pixels wide, while most of today's systems support at least
Full HD (1920x1080).
To fix this issue this change adjusts the ebuild to automatically resize
the picture before installing it. We could also do that as an offline
step, but in that case we have burden of making the result accessible
somewhere.