Install python-no-venv script for dev-python/pyvenv-auto
There are cases where virtual environment get in the way of meaningful
functionality. For example, when a Vim plugin or Git commit hook is
written in Python and it is (indirectly) invoked from within a virtual
environment, chances are that it will not work, because the required
packages are not installed in the virtual environment.
To provide an escape hatch for this problem, this change adjusts the
ebuild to install a script, python-no-venv, that forwards to the real
Python interpreter but disables any active virtual environments first.