Update dev-lang/rust-bin ebuild to 20180629
This change updates the dev-lang/rust-bin ebuild to 20180629. Note that
we remove the old ebuild because the downloaded archive does not contain
a version or other identifier and causes a clash in the manifest,
causing us to be able to only have a single version of the
dev-lang/rust-bin ebuild available at any time.