Remove no longer used dev-util/cargo-ebuild overlay
This change removes the dev-util/cargo-ebuild overlay. The overlay was
used because neither the official Portage tree nor the Rust overlay did
not contain an ebuild for the cargo-ebuild program. By now, however, a
version of this program is present in the Rust overlay in the form of
app-portage/cargo-ebuild.