Remove no longer used dev-util/clippy ebuilds
The dev-util/clippy ebuild is very particular when it comes to which
version of Rust to build with. Until recently we were only able to have
one dev-lang/rust-bin ebuild available and with the recent upgrade of
clippy to 0.0.212 the rust-bin ebuild required got bumped. With that
bump, the old versions are no longer available which also means that
older clippy versions will no longer be able to build.
To that end this change removes said old versions of the overlay.