Add dev-util/rustfmt-nightly ebuild
This change adds a new ebuild for rustfmt-nightly. This program replaces
rustfmt. So far we tried to avoid switching to rustfmt-nightly because
of breaking changes both in terms of the configuration file properties
understood as well as the formatted code produced.
However, we have to upgrade at some point because the old rustfmt simply
will not work well with newer versions of Rust. For example, the dyn
trait syntax is not understood at all, causing errors during formatting
on code that uses it.