Move dev-util/cargo patch to dev-lang/rust
There are two ways to install cargo. It can be installed directly
through the dev-util/cargo ebuild or it can be pulled in via the 'cargo'
USE flag for dev-lang/rust. In the future we would like to install
Rust's Language Server (rls). It requires the cargo during build. As the
rls is also part of dev-lang/rust (enabled through a USE flag) and
because we would like to have our cargo patch in effect for both the rls
as well as cargo itself, it makes sense for us to consume it through
dev-lang/rust, because then we can get away with applying only a single
patch.
In preparation for this step, this change moves the dev-util/cargo patch
adding support for the CARGO_TARGET_DIR_PREFIX environment variable to
dev-lang/rust.