Add dev-lang/rust patch to make rls use racer as a fallback
The rls seems to have problems looking up the definitions of symbols in
the standard library. For racer we have installed the std lib sources
and set RUST_SRC_PATH to point to them.
With this change we add a patch to adjust the rls to fall back to using
racer for definition lookups unless explicitly changed.