Remove no longer used dev-util/racer overlay
We recently switched to using the Rust Language Server for code
completion and similar tasks. While it has a fallback to using racer
(for example, for symbols present in the standard library), it actually
has the relevant functionality built in.
Hence, we no longer require the dev-util/racer overlay at all. This
change removes it.