Use gentoo overlay of dev-lang/rust as baseline for our overlay
The earlier introduction of an overlay for the dev-lang/rust ebuild was
derived from the ebuild as taken from the semi-official rust overlay.
That was done mostly because said overlay is typically faster to update
to a new version of Rust onced it is released. However, this overlay's
ebuild does not support the 'clippy' USE flag. As we really would like
to use a clippy which works with the compiler we use, this change uses
the gentoo overlay's ebuild as the baseline for our ebuild.