Add sys-devel/gdb patch to dynamically source init files
In order to customize gdb somewhat more and from different ebuilds we
would like to dynamically load all configuration files in a certain
directoy. Unfortunately, gdb seemingly cannot do that out of the box.
Although it supports a global configuration file if the
--with-system-gdbinit configure option has been provided (which is not
done by default), it is unclear how to source all files based on a
pattern.
Similarly, the --with-gdb-datadir configure option supposedly does
something along those lines as well, but we never got it working.

Instead, with this change we add a patch that just modifies the Python
init file to dynamically find and source all files in the
/usr/share/gdb/ directory.