Adjust dev-lang/rust ebuild to include better rust-gdb support
The dev-lang/rust ebuild comes with a wrapper script for GDB that sets
up things to make the debugging experience more pleasant. However, by
default there is no source code available for the standard library. That
is a problem because it makes debugging a pain. Since we already provide
the dev-lang/rust-sources ebuild that installs those sources, we should
be able to use them.
However, by default debug information will (correctly) reference the
path of the standard library it is built against. Since that standard
library comes from the rustc package and is just used from it, the paths
will reference some temporary directory used during build.
Hence, with this change we add a GDB initialization file that contains a
'substitute-path' command which is meant to be used for exactly those
cases where the path to the source has changed.