Remove dev-lang/rust-sources overlay
As it turns out the dev-lang/rust ebuild installs the Rust source tree
on its own. Whether that is a recent change or has been the case for a
while is unknown, but that means that the dev-lang/rust-sources ebuild
is no longer needed.
To that end, this change removes it while adjusting the dev-lang/rust
ebuild to set the respective environment variable to let racer et al.
work properly.