Move no longer required app-misc/tmux patch
Version 2.9a of app-misc/tmux no longer build due to problems applying
the wrap-around patch. Experiments seem to indicate that this version of
tmux actually no longer has the issues that the patch addresses and so
with this change we move the patch into a version specific subfolder,
causing it to no longer be applied for most recent versions of the
program.