Update dev-lang/rust cargo patch to work properly with workspaces
This change updates our patch for honoring the CARGO_TARGET_DIR_PREFIX
environment variable to work properly in the prospect of workspaces. The
problem the previous version had was that when individual members (i.e.,
crates) from the workspace were built, they ended up not sharing common
dependencies because their build artifacts were stored in different
directories that depended on the crates' full paths and not that of the
workspace.