Move no longer required app-misc/tmux patch
Version 3.0a of app-misc/tmux no longer builds due to problems applying
the automatic window renaming patch. Experiments seem to indicate that
this version of the program actually no longer has the issues that the
patch addresses and so with this change we move the patch into a version
specific subfolder, causing it to no longer be applied for the most
recent versions of the program.