Add sys-apps/portage patch to emit double quote entries in profile.env
The XDG specification, honored by various programs, provides a way for
programs to know where to store data of a certain type. It also provides
users with the ability to change paths.
One such approach to changing is the ~/.config/user-dirs.dirs file,
which can contain overwrites for various variables.
Unfortunately, this file cannot be used to configure the cache
directory, represented by XDG_CACHE_HOME. To that end, what we ended up
doing was to set this value in /etc/env.d/30xdg-data-local. The
problem with that is that now we have a global cache directory. That
can't work because of permission issues (root will write into it and
then nobody else can touch it).
So we defined it as follows:
  > XDG_CACHE_HOME="/tmp/.cache-${USER}"

The problem with that is that env-update, which generates, among other
things, /etc/profile.env, will write variable exports that are single
quoted. Hence, the ${USER} part will be taken literally. Due to a lack
of alternatives for solving the initial issue of setting XDG_CACHE_HOME,
this change add a patch that adjusts Portage's env-update to emit double
quoted definitions instead.