Add development snapshot ebuild for dev-util/cargo-ebuild
Both the latest crates.io released as well as Portage tree supplied
version of dev-util/cargo-ebuild are broken. With this change we add an
ebuild that retrieves the latest development snapshot that is known to
be okay-ish. It furthermore adds a patch that fixes a regression that
has existed since version 0.3.1 of the program.