Remove fish USE flag from app-shells/aliases ebuild
This change removes the fish USE flag from the app-shells/aliases
ebuild. Apparently, Gentoo has a policy against the usage of USE flags
for files that are optional for use. In any event, given that we only
use the fish shell and want the corresponding functionality basically
unconditionally, this change removes the USE flag from the ebuild.