Pin dev-lang/rust-analyzer to a known good version
Having a live version for a project such as rust-analyzer that is very
sensitive to all sorts of issues is a recipe for disaster: regressions
can happen easily and divergences between different systems are hard to
track down.
To fix these issues, this change pins the dev-lang/rust-analyzer ebuild
to a specific revision, which we bump as time passes on.