Fix www-client/chromium ebuild to make tab moving work again
This change fixes a bug in our www-client/chromium patch that adjusts
key bindings that has snug in with a recent update: upstream has decided
to use the same keyboard shortcut that we patch in earlier, rendering
our definition useless. To fix the issue, we just remove the upstream
definition.