Introduce no-clowntown Python eclasses
With the switch to Python 3.10, we are *once again* facing the ordeal of
updating all our Python based ebuilds to have updated PYTHON_COMPAT
strings. This is not only braindead work, it is also dumb at a
conceptual level: if an ebuild supports using Python 3.9 it also
supports using Python 3.10, as that is what the minor versioning scheme
is meant to express. Now there surely is some questionable reason that
justifies this braindeadness upstream, but there is absolutely no reason
to accept this work for my ebuilds.
To that end, this change introduces no-clowntown versions of the
python-r1 and python-single-r1 eclasses that sets the PYTHON_COMPAT to
all "known good" CPython versions. All ebuilds that use it can be
updated by merely updating the corresponding eclass, instead of each one
in isolation.