Update www-client/chromium patch with new key binding
In the past some existing functionality occupied the Ctrl-L key
combination that is traditionally used for selecting the location bar
and so we had to bind that functionality to Ctrl-E instead.
It appears now that this shortcut has gotten freed up and so with this
change we adjust our patch to use it for this very purpose.