Add app-eselect/eselect-rust patch to not remove symlinks
We want different Rust providers to provide different parts of the
toolchain and allow for concurrent use. Specifically, we want a
cargo-fmt coming from dev-lang/rust-bin (nightly) while dev-lang/rust is
active.
Currently, app-eselect/eselect-rust makes sure to remove symbolic links
unconditionally when providers are switched. With this change we add a
patch reworking this logic to get the desired behavior.