Clean up app-shells/fzy patches
Somehow we ended up with app-text/fzy and app-shells/fzy overlays, both
for the same program. That caused some confusion with respect to the
custom patches that we carry around, resulting in old bugs manifesting
again.
This change moves the fzy-02-fzy-tmux-use-named-pipe-for-conveying-exit-code.patch
patch previously applied to app-shells/fzy over to the app-text/fzy
category and removes app-shells/fzy altogether.